% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_politics_col.R
\name{add_politics_col}
\alias{add_politics_col}
\title{Add the political orientation from the analysis data set to the redraw data set.}
\usage{
add_politics_col(analysis_data, redraw_data)
}
\arguments{
\item{analysis_data}{The analysis data set containing the relevant demographic information.}

\item{redraw_data}{The redraw data set from participants the memory recall task.}
}
\value{
The redraw data set with the participants political orientation in the first column.
}
\description{
Add the political orientation from the analysis data set to the redraw data set.
}
