% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_salient_col.R
\name{add_salient_col}
\alias{add_salient_col}
\title{Add a column that denotes salience.}
\usage{
add_salient_col(df)
}
\arguments{
\item{df}{The data frame from the make_clean_redraw function.}
}
\value{
A data frame with a salience column - to be used for the statistical analysis.
}
\description{
Add a column that denotes salience.
}
