% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_analysis.R
\name{analysis_csv}
\alias{analysis_csv}
\title{Read in the analysis csv data frame without unused columns}
\usage{
analysis_csv(path2)
}
\arguments{
\item{path2}{The string variable that is passed from the clean_analysis function. The analysis data file.}
}
\value{
The analysis data frame without columns that end with signbefore and donatebefore
}
\description{
Read in the analysis csv data frame without unused columns
}
