% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redraw_data.R
\name{change_col_names}
\alias{change_col_names}
\title{Iterate over the data frame and rename the default R columns so that it is named only x or y.}
\usage{
change_col_names(data)
}
\arguments{
\item{data}{The over_1000_data variable.}
}
\value{
A data frame that with the 5 unique column names and no values are less then 1000
}
\description{
Iterate over the data frame and rename the default R columns so that it is named only x or y.
}
