% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_analysis.R
\name{clean_analysis}
\alias{clean_analysis}
\title{Tidy and clean data for plotting and analysis}
\usage{
clean_analysis(path1, path2, question)
}
\arguments{
\item{path1}{A string variable of the data file path for the mturk data set.}

\item{path2}{A string variable for the data file path for the analysis data set.}

\item{question}{A string variable of the attention check question.}
}
\value{
A clean tibble that is ready for more specific data cleaning (plotting and statistical analysis).
}
\description{
Tidy and clean data for plotting and analysis
}
