% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_for_plot.R
\name{clean_for_plot}
\alias{clean_for_plot}
\title{Create the y mean column from the grouped_by poli and x.}
\usage{
clean_for_plot(df)
}
\arguments{
\item{df}{The data set from the make_clean_redraw function.}
}
\value{
A clean data set ready to create the redraw plots.
}
\description{
Create the y mean column from the grouped_by poli and x.
}
