% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw_clean_data.R
\name{clean_redraw}
\alias{clean_redraw}
\title{Cleans the redraw data set a bit.}
\usage{
clean_redraw(redraw, discard)
}
\arguments{
\item{redraw}{The resulting data frame from the redraw_csv function.}

\item{discard}{The resulting vector from the discard_df function.}
}
\value{
a clean data set with the columns subject, MTurk, X7:X726
}
\description{
Cleans the redraw data set a bit.
}
