% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_analysis.R
\name{drop_failed_attention_check}
\alias{drop_failed_attention_check}
\title{Find out which particpants failed the attention check}
\usage{
drop_failed_attention_check(question, data)
}
\arguments{
\item{question}{The string variable of the rejection question.}

\item{data}{The analysis data frame that contains participants' response data}
}
\value{
A numeric vector that contains particpant numbers that failed the attention check.
}
\description{
Find out which particpants failed the attention check
}
