% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_analysis.R
\name{filter_petition}
\alias{filter_petition}
\title{Using the numeric vector, filter out participants that failed to pass the attention check. As well, exclude other known instances of non-participant data.}
\usage{
filter_petition(data, reject)
}
\arguments{
\item{data}{The analysis data set.}

\item{reject}{The numeric vector created by the drop_failed_attention_check function}
}
\value{
A clean data frame that only contains data from particpants that passed the attention check.
}
\description{
Using the numeric vector, filter out participants that failed to pass the attention check. As well, exclude other known instances of non-participant data.
}
