% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_plot.R
\name{flat_plot}
\alias{flat_plot}
\title{Makes the plot where the salient column is <= the mid point.}
\usage{
flat_plot(df, image, mid)
}
\arguments{
\item{df}{The data set created from the clean_for_plot function.}

\item{image}{A string with the image path.}

\item{mid}{The mid point: 366}
}
\value{
A ggplot graph with the first half of the data aka the flat portion of the data.
}
\description{
Makes the plot where the salient column is <= the mid point.
}
