% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_redraw.R
\name{make_clean_redraw}
\alias{make_clean_redraw}
\title{Pivot the data a bit}
\usage{
make_clean_redraw(data)
}
\arguments{
\item{data}{The data frame produced by tidy_redraw_data function.}
}
\value{
A clean and tidy data set that is ready for plotting and analysis.
}
\description{
Pivot the data a bit
}
