% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_redraw.R
\name{make_longer}
\alias{make_longer}
\title{Make the data frame longer.}
\usage{
make_longer(data)
}
\arguments{
\item{data}{The data produced by the tidy_redraw_function aka the data provided by the user.}
}
\value{
A long form data set.
}
\description{
Make the data frame longer.
}
