% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_redraw.R
\name{make_wider}
\alias{make_wider}
\title{Take the now too long data set and make it perfectly tidy for future analysis.}
\usage{
make_wider(data)
}
\arguments{
\item{data}{The data produced bt the make_longer function.}
}
\value{
A nice tidy data ready for plotting and analysis - with some minor changes.
}
\description{
Take the now too long data set and make it perfectly tidy for future analysis.
}
