% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_analysis.R
\name{mturk_csv}
\alias{mturk_csv}
\title{Read in the mturk csv data file and drop all NAs}
\usage{
mturk_csv(path1)
}
\arguments{
\item{path1}{The string variable that is passed on from the clean_analysis function. The mturk data file.}
}
\value{
The mturk data frame without NA values
}
\description{
Read in the mturk csv data file and drop all NAs
}
