% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_politics_col.R
\name{mturk_poli_data}
\alias{mturk_poli_data}
\title{Find the political orientation of each participant.}
\usage{
mturk_poli_data(data)
}
\arguments{
\item{data}{The analysis data set - that which contains all the relevant demographic information.}
}
\value{
A data frame of the analysis grouped by participants with political orientation information.
}
\description{
Find the political orientation of each participant.
}
