% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_plot.R
\name{plot_no_line}
\alias{plot_no_line}
\title{Graph without the geom_line geometric obeject.}
\usage{
plot_no_line(df, image)
}
\arguments{
\item{df}{The data frame from the clean_from_plot function.}

\item{image}{The string image path.}
}
\description{
Graph without the geom_line geometric obeject.
}
