% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_plot.R
\name{plot_with_line}
\alias{plot_with_line}
\title{Graph with the geom_line geometric.}
\usage{
plot_with_line(df, image)
}
\arguments{
\item{df}{The data frame from the clean_from_plot function.}

\item{image}{The string image path.}
}
\description{
Graph with the geom_line geometric.
}
