% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw_clean_data.R
\name{redraw_clean_data}
\alias{redraw_clean_data}
\title{Data cleaning for the participant redraw data set}
\usage{
redraw_clean_data(path, discard_data)
}
\arguments{
\item{path}{A string variable for path for the redraw data set.}

\item{discard_data}{A numeric vector of the unique participant numbers}
}
\value{
A data frame containing the redraw data but cleaned up a bit.
}
\description{
Data cleaning for the participant redraw data set
}
