% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw_clean_data.R
\name{redraw_csv}
\alias{redraw_csv}
\title{Read in the redraw data with no column names}
\usage{
redraw_csv(path)
}
\arguments{
\item{path}{A string variable of the path for the redraw data set.}
}
\value{
A data frame of the redraw data set with default R columns.
}
\description{
Read in the redraw data with no column names
}
