% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw_clean_data.R
\name{redraw_discard}
\alias{redraw_discard}
\title{Filter out participants that are not in the petition data set}
\usage{
redraw_discard(discard_data)
}
\arguments{
\item{discard_data}{}
}
\value{
A vector of the unique participant numbers from the petition data set.
}
\description{
Filter out participants that are not in the petition data set
}
