% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_politics_col.R
\name{relocate_data}
\alias{relocate_data}
\title{Relocate the the poli column to be the first column of the data set.}
\usage{
relocate_data(data)
}
\arguments{
\item{data}{The data frame produced by the left_join_data function.}
}
\value{
Exactly the same data frame as the df produced from left_join_data but the poli column has been relocated.
}
\description{
Relocate the the poli column to be the first column of the data set.
}
