% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rise_data.R
\name{rise_data}
\alias{rise_data}
\title{Clean and manipulate the data from add_salient_col to be used for statistical analysis.}
\usage{
rise_data(df)
}
\arguments{
\item{df}{The data set provided by the make_clean_redraw function.}
}
\value{
A data set with only rise data that is ready for statistical tests.
}
\description{
Clean and manipulate the data from add_salient_col to be used for statistical analysis.
}
