% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_redraw_data.R
\name{tidy_redraw_data}
\alias{tidy_redraw_data}
\title{Turn the redraw data with political orientation into something that is tidy}
\usage{
tidy_redraw_data(data)
}
\arguments{
\item{data}{The data from produced by add_politics_col function}
}
\value{
A data frame that has 5 unique column names: poli, subject, MTurk, x, and y.
}
\description{
Turn the redraw data with political orientation into something that is tidy
}
