% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_politics_col.R
\name{unique_mturk_poli}
\alias{unique_mturk_poli}
\title{Get the political orientation for each unique participant via the MTurk code.}
\usage{
unique_mturk_poli(data)
}
\arguments{
\item{data}{The data set created by the mturk_poli_data function.}
}
\value{
A data frame with the unique MTurk code and the corresponding political orientation.
}
\description{
Get the political orientation for each unique participant via the MTurk code.
}
