% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_plot.R
\name{whole_plot}
\alias{whole_plot}
\title{Create the average redrawn plot by political orientation.}
\usage{
whole_plot(img_path, df, line = FALSE)
}
\arguments{
\item{img_path}{A string that has the path for the original plot image.}

\item{df}{The resulting data frame from the clean_for_plot function.}

\item{line}{Default set to false, add geom_line grouped by political orientation if set to true.}
}
\value{
A ggplot graph with the recreated redraw map from the recall task.
}
\description{
Create the average redrawn plot by political orientation.
}
